Cardano.Ledger.Dijkstra.PParams

Documentation

data DijkstraPParams f era

Dijkstra Protocol parameters. The following parameters have been added since Dijkstra: * maxRefScriptSizePerBlock * maxRefScriptSizePerTx * refScriptCostStride * refScriptCostMultiplier

Constructors

DijkstraPParams

Fields

dppMinFeeA ∷ !(THKD ('PPGroups 'EconomicGroup 'SecurityGroup) f Coin)

The linear factor for the minimum fee calculation

dppMinFeeB ∷ !(THKD ('PPGroups 'EconomicGroup 'SecurityGroup) f Coin)

The constant factor for the minimum fee calculation

dppMaxBBSize ∷ !(THKD ('PPGroups 'NetworkGroup 'SecurityGroup) f Word32)

Maximal block body size

dppMaxTxSize ∷ !(THKD ('PPGroups 'NetworkGroup 'SecurityGroup) f Word32)

Maximal transaction size

dppMaxBHSize ∷ !(THKD ('PPGroups 'NetworkGroup 'SecurityGroup) f Word16)

Maximal block header size

dppKeyDeposit ∷ !(THKD ('PPGroups 'EconomicGroup 'NoStakePoolGroup) f Coin)

The amount of a key registration deposit

dppPoolDeposit ∷ !(THKD ('PPGroups 'EconomicGroup 'NoStakePoolGroup) f (CompactForm Coin))

The amount of a pool registration deposit

dppEMax ∷ !(THKD ('PPGroups 'TechnicalGroup 'NoStakePoolGroup) f EpochInterval)

Maximum number of epochs in the future a pool retirement is allowed to be scheduled for.

dppNOpt ∷ !(THKD ('PPGroups 'TechnicalGroup 'NoStakePoolGroup) f Word16)

Desired number of pools

dppA0 ∷ !(THKD ('PPGroups 'TechnicalGroup 'NoStakePoolGroup) f NonNegativeInterval)

Pool influence

dppRho ∷ !(THKD ('PPGroups 'EconomicGroup 'NoStakePoolGroup) f UnitInterval)

Monetary expansion

dppTau ∷ !(THKD ('PPGroups 'EconomicGroup 'NoStakePoolGroup) f UnitInterval)

Treasury expansion

dppProtocolVersion ∷ !(HKDNoUpdate f ProtVer)

Protocol version

dppMinPoolCost ∷ !(THKD ('PPGroups 'EconomicGroup 'NoStakePoolGroup) f Coin)

Minimum Stake Pool Cost

dppCoinsPerUTxOByte ∷ !(THKD ('PPGroups 'EconomicGroup 'SecurityGroup) f CoinPerByte)

Cost in lovelace per byte of UTxO storage

dppCostModels ∷ !(THKD ('PPGroups 'TechnicalGroup 'NoStakePoolGroup) f CostModels)

Cost models for non-native script languages

dppPrices ∷ !(THKD ('PPGroups 'EconomicGroup 'NoStakePoolGroup) f Prices)

Prices of execution units (for non-native script languages)

dppMaxTxExUnits ∷ !(THKD ('PPGroups 'NetworkGroup 'NoStakePoolGroup) f OrdExUnits)

Max total script execution resources units allowed per tx

dppMaxBlockExUnits ∷ !(THKD ('PPGroups 'NetworkGroup 'SecurityGroup) f OrdExUnits)

Max total script execution resources units allowed per block

dppMaxValSize ∷ !(THKD ('PPGroups 'NetworkGroup 'SecurityGroup) f Word32)

Max size of a Value in an output

dppCollateralPercentage ∷ !(THKD ('PPGroups 'TechnicalGroup 'NoStakePoolGroup) f Word16)

Percentage of the txfee which must be provided as collateral when including non-native scripts.

dppMaxCollateralInputs ∷ !(THKD ('PPGroups 'NetworkGroup 'NoStakePoolGroup) f Word16)

Maximum number of collateral inputs allowed in a transaction

dppPoolVotingThresholds ∷ !(THKD ('PPGroups 'GovGroup 'NoStakePoolGroup) f PoolVotingThresholds)

Thresholds for SPO votes

dppDRepVotingThresholds ∷ !(THKD ('PPGroups 'GovGroup 'NoStakePoolGroup) f DRepVotingThresholds)

Thresholds for DRep votes

dppCommitteeMinSize ∷ !(THKD ('PPGroups 'GovGroup 'NoStakePoolGroup) f Word16)

Minimum size of the Constitutional Committee

dppCommitteeMaxTermLength ∷ !(THKD ('PPGroups 'GovGroup 'NoStakePoolGroup) f EpochInterval)

The Constitutional Committee Term limit in number of Slots

dppGovActionLifetime ∷ !(THKD ('PPGroups 'GovGroup 'NoStakePoolGroup) f EpochInterval)

Gov action lifetime in number of Epochs

dppGovActionDeposit ∷ !(THKD ('PPGroups 'GovGroup 'SecurityGroup) f Coin)

The amount of the Gov Action deposit

dppDRepDeposit ∷ !(THKD ('PPGroups 'GovGroup 'NoStakePoolGroup) f (CompactForm Coin))

The amount of a DRep registration deposit

dppDRepActivity ∷ !(THKD ('PPGroups 'GovGroup 'NoStakePoolGroup) f EpochInterval)

The number of Epochs that a DRep can perform no activity without losing their Active status.

dppMinFeeRefScriptCostPerByte ∷ !(THKD ('PPGroups 'EconomicGroup 'SecurityGroup) f NonNegativeInterval)

Reference scripts fee for the minimum fee calculation TODO ensure that the groups here make sense

dppMaxRefScriptSizePerBlock ∷ !(THKD ('PPGroups 'NetworkGroup 'SecurityGroup) f Word32)

Limit on the total number of bytes of all reference scripts combined from all transactions within a block.

dppMaxRefScriptSizePerTx ∷ !(THKD ('PPGroups 'NetworkGroup 'SecurityGroup) f Word32)

Limit on the total number of bytes of reference scripts that a transaction can use.

dppRefScriptCostStride ∷ !(THKD ('PPGroups 'NetworkGroup 'SecurityGroup) f (NonZero Word32))

dppRefScriptCostMultiplier ∷ !(THKD ('PPGroups 'NetworkGroup 'SecurityGroup) f PositiveInterval)

class DijkstraEraPParams era ⇒ DijkstraEraPParams era where

Methods

hkdMaxRefScriptSizePerBlockL ∷ Lens' (PParamsHKD f era) (HKD f Word32)

hkdMaxRefScriptSizePerTxL ∷ Lens' (PParamsHKD f era) (HKD f Word32)

hkdRefScriptCostStrideL ∷ Lens' (PParamsHKD f era) (HKD f (NonZero Word32))

hkdRefScriptCostMultiplierL ∷ Lens' (PParamsHKD f era) (HKD f PositiveInterval)

data UpgradeDijkstraPParams f era

Constructors

UpgradeDijkstraPParams

Fields

udppMaxRefScriptSizePerBlock ∷ !(HKD f Word32)

udppMaxRefScriptSizePerTx ∷ !(HKD f Word32)

udppRefScriptCostStride ∷ !(HKD f (NonZero Word32))

udppRefScriptCostMultiplier ∷ !(HKD f PositiveInterval)

ppRefScriptCostMultiplierL ∷ DijkstraEraPParams era ⇒ Lens' (PParams era) PositiveInterval

ppRefScriptCostStrideL ∷ DijkstraEraPParams era ⇒ Lens' (PParams era) (NonZero Word32)

ppMaxRefScriptSizePerTxL ∷ DijkstraEraPParams era ⇒ Lens' (PParams era) Word32

ppMaxRefScriptSizePerBlockL ∷ DijkstraEraPParams era ⇒ Lens' (PParams era) Word32

ppuRefScriptCostMultiplierL ∷ DijkstraEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe PositiveInterval)

ppuRefScriptCostStrideL ∷ DijkstraEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe (NonZero Word32))

ppuMaxRefScriptSizePerTxL ∷ DijkstraEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Word32)

ppuMaxRefScriptSizePerBlockL ∷ DijkstraEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Word32)

Orphan instances

AlonzoEraPParams DijkstraEra

Instance details

Methods

hkdCoinsPerUTxOWordL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, ExactEra AlonzoEra DijkstraEra) ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f CoinPerWord)

hkdCostModelsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f CostModels)

hkdPricesL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f Prices)

hkdMaxTxExUnitsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f ExUnits)

hkdMaxBlockExUnitsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f ExUnits)

hkdMaxValSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f Natural)

hkdCollateralPercentageL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f Natural)

hkdMaxCollateralInputsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f Natural)

BabbageEraPParams DijkstraEra

Instance details

Methods

hkdCoinsPerUTxOByteL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f CoinPerByte)

ConwayEraPParams DijkstraEra

Instance details

Methods

modifiedPPGroups ∷ PParamsUpdate DijkstraEra → Set PPGroups

ppuWellFormed ∷ ProtVer → PParamsUpdate DijkstraEra → Bool

hkdPoolVotingThresholdsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f PoolVotingThresholds)

hkdDRepVotingThresholdsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f DRepVotingThresholds)

hkdCommitteeMinSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f Natural)

hkdCommitteeMaxTermLengthL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f EpochInterval)

hkdGovActionLifetimeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f EpochInterval)

hkdGovActionDepositL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f Coin)

hkdDRepDepositCompactL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f (CompactForm Coin))

hkdDRepActivityL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f EpochInterval)

hkdMinFeeRefScriptCostPerByteL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f NonNegativeInterval)

ppMaxRefScriptSizePerTxG ∷ SimpleGetter (PParams DijkstraEra) Word32

ppMaxRefScriptSizePerBlockG ∷ SimpleGetter (PParams DijkstraEra) Word32

ppRefScriptCostMultiplierG ∷ SimpleGetter (PParams DijkstraEra) PositiveInterval

ppRefScriptCostStrideG ∷ SimpleGetter (PParams DijkstraEra) (NonZero Word32)

EraPParams DijkstraEra

Instance details

Associated Types

type PParamsHKD f DijkstraEra = (r ∷ Type)

type UpgradePParams f DijkstraEra

type DowngradePParams f DijkstraEra

Methods

applyPPUpdates ∷ PParams DijkstraEra → PParamsUpdate DijkstraEra → PParams DijkstraEra

emptyPParamsIdentity ∷ PParamsHKD Identity DijkstraEra

emptyPParamsStrictMaybe ∷ PParamsHKD StrictMaybe DijkstraEra

emptyUpgradePParamsUpdate ∷ UpgradePParams StrictMaybe DijkstraEra

upgradePParamsHKD ∷ ∀ (f ∷ Type → Type). (HKDApplicative f, EraPParams (PreviousEra DijkstraEra)) ⇒ UpgradePParams f DijkstraEra → PParamsHKD f (PreviousEra DijkstraEra) → PParamsHKD f DijkstraEra

downgradePParamsHKD ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, EraPParams (PreviousEra DijkstraEra)) ⇒ DowngradePParams f DijkstraEra → PParamsHKD f DijkstraEra → PParamsHKD f (PreviousEra DijkstraEra)

hkdMinFeeAL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f Coin)

hkdMinFeeBL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f Coin)

hkdMaxBBSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f Word32)

hkdMaxTxSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f Word32)

hkdMaxBHSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f Word16)

hkdKeyDepositL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f Coin)

hkdPoolDepositCompactL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f (CompactForm Coin))

hkdEMaxL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f EpochInterval)

hkdNOptL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f Word16)

hkdA0L ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f NonNegativeInterval)

hkdRhoL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f UnitInterval)

hkdTauL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f UnitInterval)

hkdDL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, AtMostEra "Alonzo" DijkstraEra) ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f UnitInterval)

ppDG ∷ SimpleGetter (PParams DijkstraEra) UnitInterval

hkdExtraEntropyL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, AtMostEra "Alonzo" DijkstraEra) ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f Nonce)

hkdProtocolVersionL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, AtMostEra "Babbage" DijkstraEra) ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f ProtVer)

ppProtocolVersionL ∷ Lens' (PParams DijkstraEra) ProtVer

ppuProtocolVersionL ∷ Lens' (PParamsUpdate DijkstraEra) (StrictMaybe ProtVer)

hkdMinUTxOValueL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, AtMostEra "Mary" DijkstraEra) ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f Coin)

hkdMinPoolCostL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f DijkstraEra) (HKD f Coin)

eraPParams ∷ [PParam DijkstraEra]